improve(isearch): disable isearch lazy count
After think for a while, I'm aware that I don't need this feature
anymore. Actually, this feature is annoying many times when I am
typing and focusing on minibuffer to check if I type correctly. and in
many times, the number of matched item is a redundant information, if
I really want to know the number of matched item, I can use `occur' by
pressing "M-s o".

So the conclusion is that I will use `isearch' to quickly move cursor
around and use `occur' to know the number of matched item and navigate.