improve(ibuffer): ibuffer-auto & hl-line finally work well together!
Before, I set two hook after I go to ibuffer(hl-line-mode &
ibuffer-auto-mode) but there is a very annoying thing. When I first
time call ibuffer/ibuffer-jump, it works as expected, there is a
highlight line and ibuffer-auto-mode is on. but when I quit ibuffer
window (not kill this buffer), the next time I call ibuffer or
ibuffer-jump, there is not a highlight line until I press any key or
call any command. I just want it show highlight line when I go to
ibuffer, now it works as expected.